+++ src/app/(main)/dashboard/patient-monitoring/_components/vital-trend-chart.tsx
@@ -1,34 +1,21 @@
-import { useMemo } from "react";
-
 import { CartesianGrid, Line, LineChart, YAxis } from "recharts";
 
 import { type ChartConfig, ChartContainer } from "@/components/ui/chart";
 import { cn } from "@/lib/utils";
 
-import type { CardiacRhythm, PatientRecord } from "./data";
-import { useTrendTick } from "./use-realtime-tick";
-import { normalizedTrendTemplate } from "./waveform-data";
-
 export type TrendKind = "heart-rate" | "map" | "spo2";
 
-interface VitalTrendChartProps {
-  kind: TrendKind;
-  patient: PatientRecord;
-}
-
-interface TrendPoint {
-  sample: number;
+export interface TrendPoint {
   value: number;
 }
 
-interface TrendDefinition {
-  baseline: "arterialMap" | "heartRate" | "spo2";
+interface VitalTrendChartProps {
+  ariaLabel: string;
+  data: TrendPoint[];
   domain: [number, number];
+  kind: TrendKind;
   ticks: number[];
-  variation: number;
 }
-
-const TREND_POINT_COUNT = 90;
 
 const trendChartConfig = {
   value: {
@@ -43,81 +30,10 @@ const trendClasses: Record<TrendKind, string> = {
   spo2: "text-cyan-500 dark:text-cyan-400",
 };
 
-const trendDefinitions = {
-  "heart-rate": {
-    baseline: "heartRate",
-    domain: [50, 150],
-    ticks: [50, 100, 150],
-    variation: 1.4,
-  },
-  map: {
-    baseline: "arterialMap",
-    domain: [50, 150],
-    ticks: [50, 100, 150],
-    variation: 4,
-  },
-  spo2: {
-    baseline: "spo2",
-    domain: [80, 100],
-    ticks: [80, 90, 100],
-    variation: 0.65,
-  },
-} satisfies Record<TrendKind, TrendDefinition>;
-
-const heartRateVariations: Record<CardiacRhythm, number> = {
-  "atrial-fibrillation": 4.5,
-  "low-voltage": 1.4,
-  "occasional-pvc": 3,
-  paced: 1,
-  sinus: 1.4,
-  "sinus-bradycardia": 1.2,
-  "sinus-tachycardia": 2,
-  "st-depression": 1.8,
-};
-
-function getTemplateValue(patient: PatientRecord, sample: number) {
-  const phaseOffset = Math.round(patient.signalProfile.phase * normalizedTrendTemplate.length);
-  const index =
-    (((sample + phaseOffset) % normalizedTrendTemplate.length) + normalizedTrendTemplate.length) %
-    normalizedTrendTemplate.length;
-  return normalizedTrendTemplate[index] ?? 0;
-}
-
-function getTrendValue(patient: PatientRecord, kind: TrendKind, sample: number) {
-  const definition = trendDefinitions[kind];
-  const templateValue = getTemplateValue(patient, sample);
-  const baseline = patient[definition.baseline];
-  let variation = definition.variation;
-
-  if (kind === "heart-rate") {
-    variation = heartRateVariations[patient.signalProfile.rhythm];
-  }
-
-  const value = baseline + templateValue * variation;
-  const [minimum, maximum] = definition.domain;
-
-  return Math.min(Math.max(value, minimum), maximum);
-}
-
-function createTrendData(patient: PatientRecord, kind: TrendKind, endSample: number) {
-  return Array.from({ length: TREND_POINT_COUNT }, (_, index): TrendPoint => {
-    const sample = endSample - (TREND_POINT_COUNT - index - 1);
-
-    return {
-      sample,
-      value: getTrendValue(patient, kind, sample),
-    };
-  });
-}
-
-export function VitalTrendChart({ kind, patient }: VitalTrendChartProps) {
-  const tick = useTrendTick();
-  const data = useMemo(() => createTrendData(patient, kind, tick), [kind, patient, tick]);
-  const definition = trendDefinitions[kind];
-
+export function VitalTrendChart({ ariaLabel, data, domain, kind, ticks }: VitalTrendChartProps) {
   return (
     <ChartContainer
-      aria-label={`${patient.bed} ${kind} trend`}
+      aria-label={ariaLabel}
       className={cn("aspect-auto h-14 w-full", trendClasses[kind])}
       config={trendChartConfig}
       initialDimension={{ width: 600, height: 56 }}
@@ -128,10 +44,10 @@ export function VitalTrendChart({ kind, patient }: VitalTrendChartProps) {
         <YAxis
           allowDataOverflow
           axisLine={false}
-          domain={definition.domain}
+          domain={domain}
           tickLine={false}
           tickMargin={4}
-          ticks={definition.ticks}
+          ticks={ticks}
           width={28}
         />
         <Line
